+++ application/config/routes.php
@@ -49,6 +49,10 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 | Examples:	my-controller/index	-> my_controller/index
 |		my-controller/my-method	-> my_controller/my_method
 */
-$route['default_controller'] = 'welcome';
+$route['default_controller'] = 'home';
+$route['stock-history'] = 'Stock_History/index';
+$route['player-portfolio'] = 'Player_Portfolio/index';
+$route['stock-history/([a-zA-Z]{3,4})'] = 'Stock_History/displayStock/$1';
+$route['player-portfolio/([a-zA-Z]+)'] = 'Player_Portfolio/displayPortfolio/$1';
 $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;