+++ application/views/base/_footer.php
@@ -1,0 +1,5 @@
+<div class="footer-copyright">
+    <div class="container">
+        Copyright &copy; 2016 Bulaleeplum.
+    </div>
+</div>