+++ application/views/base/_header.php
@@ -1,0 +1,8 @@
+<!-- Navigation  -->
+<nav class="cyan darken-2" role="navigation">
+    <div class="nav-wrapper container">
+        <a id="logo-container" href="/" class="brand-logo">Stock Ticker</a>
+        {logged_in_status}
+        <a href="#" data-activates="nav-mobile" class="button-collapse"><i class="material-icons">menu</i></a>
+    </div>
+</nav>