+++ assets/js/functions.js
@@ -1,0 +1,6 @@
+$( document ).ready(function(){
+    $('.button-collapse').sideNav();    // for mobile navigation menu
+    $('select').material_select();      // for <select> tags
+    $('.modal-trigger').leanModal();    // for modals boxes
+    $('.parallax').parallax();          // for the parallax sections
+});