+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Change Log
-Changes to the lab will be detailed in this log file.
+Changes to the project will be detailed in this log file.
 
 #### Feb 07 2016
 - project started
+
+#### Feb 12 2016
+- configured base controller and master template
+- created Home view/controller with dummy navbar
+
+#### Feb 15 2016
+- added materialize bootstrap and css framework
+- added a basic page for stock history
+
+#### Feb 16 2016
+- added a basic page for player portfolio
+- added SQL dump of database in root directory
+- added basic db queries in StocksModel
+- updated config to autoload StocksModel
+- player names loaded into Player Portfolio dropdown
+- commented Player Portfolio controller
+
+#### Feb 17 2016
+- Added stock history query
+- Completed table generation
+
+#### Feb 18 2016
+- Added default stock display
+- Added player portfolio query
+- Separated player portfolio table into Trading Activity and Current Holdings
+
+#### Feb 19 2015
+- Added Collections table to stockerticker database
+- Added equity and net worth queries for a player
+- Implemented routing on stock-history page
+- Refactored models
+- Implemented routing on player-portfolio page
+- Set temporary default player-portfolio view to be blank
+- Implemented stock dropdown generation
+- Fixed dropdown generation
+- Implemented initial homepage view with equity
+- Implemented user login
+- Cleanup of old files/debug statements
+- Hide portfolios and display message to user if not logged in
+- Styled pages